+++ app/src/lib/highlighter/types.ts
